菜品模块完结

+++ sky-take-out/sky-server/src/main/java/com/sky/service/impl/DishServiceImpl.java
@@ -2,12 +2,18 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
+import com.sky.constant.StatusConstant;
 import com.sky.dto.DishDTO;
 import com.sky.dto.DishPageQueryDTO;
 import com.sky.entity.Dish;
 import com.sky.entity.DishFlavor;
+import com.sky.entity.Setmeal;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.DishFlavorMapper;
 import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealDishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.DishService;
 import com.sky.vo.DishVO;
@@ -32,6 +38,10 @@ public class DishServiceImpl implements DishService {
     DishMapper dishMapper;
     @Autowired
     DishFlavorMapper dishFlavorMapper;
+    @Autowired
+    SetmealMapper setmealMapper;
+    @Autowired
+    SetmealDishMapper setmealDishMapper;
     /*
     * 新增菜品以及保存菜品相关的口味
     * */
@@ -75,9 +85,29 @@ public class DishServiceImpl implements DishService {
     /*
     * 批量删除菜品
     * */
+    @Transactional
     @Override
     public void deleteByIds(Long[] ids) {
+        //先判断当前要删除的菜品是否为启售中
+        for (Long id:ids){
+            Dish dish=dishMapper.selectById(id);
+            if (dish.getStatus()== StatusConstant.ENABLE){
+                throw new DeletionNotAllowedException(MessageConstant.DISH_ON_SALE);
+            }
+        }
+
+        //再判断当前要删除的菜品是否被套餐关联了
+        List<Long> setmealIds= setmealMapper.getSetmealIdsByDishIds(ids);
+        if (setmealIds!=null&&setmealIds.size()!=0){
+            //如果关联了，抛出业务异常
+            throw new DeletionNotAllowedException(MessageConstant.DISH_BE_RELATED_BY_SETMEAL);
+        }
+
         dishMapper.deleteByIds(ids);
+        for(Long id:ids){
+            dishFlavorMapper.delete(id);
+        }
+
     }
 
     /*
@@ -91,5 +121,56 @@ public class DishServiceImpl implements DishService {
         BeanUtils.copyProperties(dish,dishVo);
         dishVo.setFlavors(dishFlavors);
         return dishVo;
+    }
+
+    @Transactional
+    @Override
+    public void update(DishDTO dishDTO) {
+        //更新菜品表
+        Dish dish=new Dish();
+        BeanUtils.copyProperties(dishDTO,dish);
+        dishMapper.update(dish);
+
+        //删除当前菜品关联的口味数据
+        dishFlavorMapper.delete(dish.getId());
+
+        //插入最新的口味数据
+        List<DishFlavor> dishFlavors=dishDTO.getFlavors();
+        if (dishFlavors!=null&&dishFlavors.size()>0){
+            dishFlavors.forEach(dishFlavor -> {
+                dishFlavor.setDishId(dish.getId());
+            });
+        }
+        dishFlavorMapper.insertBatch(dishFlavors);
+
+
+    }
+
+    /*
+    * 更新起售或者停售
+    * */
+    @Override
+    public void updateStatus(DishDTO dishDTO) {
+        //先更新菜品表
+        Dish dish=new Dish();
+        BeanUtils.copyProperties(dishDTO,dish);
+        dishMapper.update(dish);
+        //然后停售包含该菜品的所有套餐
+        if (dish.getStatus()==StatusConstant.DISABLE){
+            Long dishId = dish.getId();
+            //根据菜品id停售相关套餐
+            List<Long> dishIds=new ArrayList<>();
+            dishIds.add(dishId);
+            List<Long> setmealIds = setmealDishMapper.getSetmealIdsByDishIds(dishIds);
+            if (setmealIds!=null&&setmealIds.size()>0){
+                for (Long setmealId:setmealIds){
+                    Setmeal setmeal = Setmeal.builder().status(StatusConstant.DISABLE).id(setmealId).build();
+                    setmealMapper.update(setmeal);
+                }
+            }
+
+        }
+
+
     }
 }
